// src/gql/graphql.ts
export type Orientation =
  | 'horizontal'
  | 'vertical';
// ...
export type PosterItemFragment = { id: string, title: string, currentPrice: number, slug: string, description?: string | null, orientation: Orientation, images: Array<{ url: string }> };